Move slicerWebWorker into main thread for compatability with phantom
