+++ src/Leaflet.VectorGrid.Slicer.js
@@ -1,69 +1,81 @@
-
-// The geojson/topojson is sliced into tiles via a web worker.
-// This import statement depends on rollup-file-as-blob, so that the
-// variable 'workerCode' is a blob URL.
-
-import workerCode from './slicerWebWorker.js.worker';
-
+import geojsonvt from 'geojson-vt';
+import * as topojson from "topojson-client";
 
 L.VectorGrid.Slicer = L.VectorGrid.extend({
 
-	options: {
-		vectorTileLayerName: 'sliced',
-		extent: 4096,	// Default for geojson-vt
-		maxZoom: 14  	// Default for geojson-vt
-	},
+  options: {
+    vectorTileLayerName: 'sliced',
+    extent: 4096,  // Default for geojson-vt
+    maxZoom: 14    // Default for geojson-vt
+  },
 
-	initialize: function(geojson, options) {
-		L.VectorGrid.prototype.initialize.call(this, options);
+  cachedOptions: {},
+  slicers: {},
 
-		// Create a shallow copy of this.options, excluding things that might
-		// be functions - we only care about topojson/geojsonvt options
-		var options = {};
-		for (var i in this.options) {
-			if (i !== 'rendererFactory' &&
-				i !== 'vectorTileLayerStyles' &&
-				typeof (this.options[i]) !== 'function'
-			) {
-				options[i] = this.options[i];
-			}
-		}
+  initialize: function(geojson, options) {
+    L.VectorGrid.prototype.initialize.call(this, options);
 
-// 		this._worker = new Worker(window.URL.createObjectURL(new Blob([workerCode])));
-		this._worker = new Worker(workerCode);
+    // Create a shallow copy of this.options, excluding things that might
+    // be functions - we only care about topojson/geojsonvt options
+    for (var i in this.options) {
+      if (i !== 'rendererFactory' &&
+        i !== 'vectorTileLayerStyles' &&
+        typeof (this.options[i]) !== 'function'
+      ) {
+        this.cachedOptions[i] = this.options[i];
+      }
+    }
 
-		// Send initial data to worker.
-		this._worker.postMessage(['slice', geojson, options]);
+    // Given a blob of GeoJSON and some topojson/geojson-vt options, configure slicers.
+    if (geojson.type && geojson.type === 'Topology') {
+      for (var layerName in geojson.objects) {
+        this.slicers[layerName] = geojsonvt(
+          topojson.feature(geojson, geojson.objects[layerName])
+        , options);
+      }
+    } else {
+      this.slicers[options.vectorTileLayerName] = geojsonvt(geojson, options);
+    }
 
-	},
+  },
 
+  _getVectorTilePromise: function(coords) {
+    var _this = this;
 
-	_getVectorTilePromise: function(coords) {
+    var p = new Promise( function (res) {
+      var tileLayers = {};
+      for (var layerName in _this.slicers) {
+        var slicedTileLayer = _this.slicers[layerName].getTile(coords.z, coords.x, coords.y);
 
-		var _this = this;
+        if (slicedTileLayer) {
+          var vectorTileLayer = {
+            features: [],
+            extent: _this.cachedOptions.extent,
+            name: _this.cachedOptions.vectorTileLayerName,
+            length: slicedTileLayer.features.length
+          }
 
-		var p = new Promise( function waitForWorker(res) {
-			_this._worker.addEventListener('message', function recv(m) {
-				if (m.data.coords &&
-				    m.data.coords.x === coords.x &&
-				    m.data.coords.y === coords.y &&
-				    m.data.coords.z === coords.z ) {
-
-					res(m.data);
-					_this._worker.removeEventListener('message', recv);
-				}
-			});
-		});
-
-		this._worker.postMessage(['get', coords]);
-
-		return p;
-	},
+          for (var i in slicedTileLayer.features) {
+            var feat = {
+              geometry: slicedTileLayer.features[i].geometry,
+              properties: slicedTileLayer.features[i].tags,
+              type: slicedTileLayer.features[i].type  // 1 = point, 2 = line, 3 = polygon
+            }
+            vectorTileLayer.features.push(feat);
+          }
+          tileLayers[layerName] = vectorTileLayer;
+        }
+      }
+      var data = { layers: tileLayers, coords: coords };
+      res(data);
+    });
+    return p;
+  },
 
 });
 
 
 L.vectorGrid.slicer = function (geojson, options) {
-	return new L.VectorGrid.Slicer(geojson, options);
+  return new L.VectorGrid.Slicer(geojson, options);
 };
